Alerts component for wizard crit page
