+++ ui/src/app/cohort-search/criteria-wizard/alerts/alerts.component.html
@@ -1,0 +1,19 @@
+<div 
+  *ngIf="hasErrors"
+  [style.padding-right]="'1rem'">
+  <clr-alert
+    [clrAlertClosable]="false"
+    [clrAlertType]="'alert-danger'">
+    <div clr-alert-item class="alert-item"
+      *ngFor="let error of criteriaErrors">
+      <span class="alert-text">
+        Failed to load children of {{ error.kind }} with ID {{ error.parentId }}.
+      </span>
+      <div class="alert-actions">
+        <a (click)="closeAlert(error)">
+          <clr-icon shape="close"></clr-icon>
+        </a>
+      </div>
+    </div>
+  </clr-alert>
+</div>
