+++ ui/src/app/cohort-search/criteria-wizard/wizard/wizard.component.css
@@ -23,10 +23,6 @@
   overflow: hidden!important;
 }
 
-.error-box {
-  padding-right: 1rem;
-}
-
 #criteria-page {
   height: 100%;
   display: flex;
